add webpack and app is divided into modules
